+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/util/Base.groovy
@@ -1,0 +1,8 @@
+package ru.vyarus.dropwizard.guice.support.util
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 22.10.2014
+ */
+class Base implements GenerifiedInterface<Integer, String[], List<String>, List<Set<String>>>{
+}