+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/util/GenerifiedInterface.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.dropwizard.guice.support.util
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 22.10.2014
+ */
+public interface GenerifiedInterface<T, K, G, Z> {
+
+}